All JS now resides in dist.min.js
This vastly improves load efficiency. dist.min.js gets generated with
Gulp. the end file is of about 450kb, however with gzip it goes down to
a little over 150kb. This makes a page like leaderboard load using
545kb, excluding the header image. Of course, with caching this goes
down to 7.2kb.

With this commit, package.json is also added, containing the new
dependencies. Also, vue has been added to the dependencies, because
that's what we're gonna try implementing next.